requirements.txt
langchain_openai 
langchain_core
langchain_community
langserve
bs4
pypdf
chromadb
faiss-cpu
groq
langchain-groq
wikipedia
arxiv
PyMuPDF
langchainhub
sentence_transformers
langchain-objectbox
langchain
torch
sentence-transformers
matplotlib